Update to match latest e164.js
This is generated by running 'rake rebuild' which updates to the latest
version from e164.js.    All changes were automatically made by that process
except for the change to unlock us from the old version of bundler and rake
which were dead and gone.   There didn't seem to be any good reason for those
versions to be locked down.